requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==1.24.4
pillow==10.4.0
proglog==0.1.10
pycolmap==0.6.1
requests==2.32.3
tqdm==4.66.4
urllib3==2.2.2
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.1.0
importlib_resources==6.4.0
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
moviepy==1.0.3
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycolmap==0.6.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tifffile==2023.7.10
tqdm==4.66.4
trimesh==4.4.3
typing_extensions==4.5.0
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
